refactor(forum-page): rename components to ui and reorganize forum blocks as organisms
