+++ packages/client/src/pages/ForumPage/ForumPage.module.css
@@ -2,11 +2,7 @@
   box-sizing: border-box;
   display: flex;
   flex-direction: column;
-
-  /* НЕ фиксируем высоту через calc */
   padding: 24px;
-
-  /* чтобы не было горизонтального скролла на зуме */
   min-width: 0;
 }
 
@@ -22,7 +18,7 @@
 
 .contentRow {
   flex: 1;
-  min-height: 0; /* ключевое для flex-скроллов */
+  min-height: 0;
   margin-top: 16px;
   align-items: stretch;
 }
@@ -35,7 +31,6 @@
   min-width: 0;
 }
 
-/* карточки должны уметь сжиматься */
 .leftCol > :global(.ant-card),
 .rightCol > :global(.ant-card) {
   flex: 1;
