+++ packages/client/src/pages/ForumPage/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './organisms'
