+++ packages/client/src/pages/ForumPage/ui/organisms/CommentsList/CommentsList.module.css
@@ -9,7 +9,6 @@
   margin: 0 0 8px;
 }
 
-/* Важно: список занимает всё доступное место и скроллится внутри */
 .list {
   --comment-row-height: 96px;
   flex: 1;
