+++ packages/client/src/pages/ForumPage/ui/organisms/CommentsList/CommentsList.tsx
@@ -1,6 +1,6 @@
 import { Avatar, Button, Empty, List, Space, Typography } from 'antd'
 import { LikeOutlined } from '@ant-design/icons'
-import { CommentItem } from '../../model/types'
+import { CommentItem } from '../../../model/types'
 import styles from './CommentsList.module.css'
 
 const { Paragraph, Text, Title } = Typography
