+++ packages/client/src/pages/ForumPage/ui/organisms/TopicList/TopicList.module.css
@@ -13,7 +13,6 @@
   min-height: 0;
 }
 
-/* Важно: список занимает всё доступное место и скроллится внутри */
 .list {
   --topic-row-height: 120px;
   flex: 1;
